Fix issue with types

+++ tokun/pipeline.py
@@ -35,7 +35,7 @@ def encode(data: tf.Tensor, token_size: int, sample_size: int, dtype: tf.dtypes.
 def binarize(data: tf.Tensor, depth: int=8, dtype: tf.dtypes.DType=None, flatten: bool=True) -> tf.Tensor:
     __dtype = dtype or data.dtype
     # big endian: most significant bit first
-    __mask = tf.bitwise.left_shift(tf.ones((), dtype=__dtype), tf.convert_to_tensor(range(depth)[::-1], dtype=__dtype))
+    __mask = tf.bitwise.left_shift(tf.ones((), dtype=tf.dtypes.int32), tf.convert_to_tensor(range(depth)[::-1], dtype=tf.dtypes.int32))
     # select each bit from the original data
     __bits = tf.bitwise.bitwise_and(tf.expand_dims(data, -1), __mask)
     # format
